move post user functions into cryptoApi
